refactor(parser-v3): Move template parsing in a separate TemplateParser
The main Parser subscribes to events of TemplateParser.
Some events are forwarded as is by calling emit(), and
some events are intercepted for additional parsing.
