+++ lib/v3/parser.js
@@ -3,7 +3,6 @@ const path = require('path');
 
 const espree = require('espree');
 const eslint = require('eslint');
-const HtmlParser = require('htmlparser2-svelte').Parser;
 
 const utils = require('./../utils');
 const jsdoc = require('./../jsdoc');
@@ -20,6 +19,9 @@ const {
     updateType
 } = require('./v3-utils');
 
+const TemplateParser = require('./template');
+const TemplateEvent = require('./template').TemplateEvent;
+
 const hasOwnProperty = utils.hasOwnProperty;
 
 const SUPPORTED_FEATURES = [
@@ -34,9 +36,12 @@ const SUPPORTED_FEATURES = [
     'slots',
     'refs'
 ];
+
 const SCOPE_DEFAULT = 'default';
 const SCOPE_STATIC = 'static';
 const SCOPE_MARKUP = 'markup';
+
+const RE_ANONYMOUS_FUNCTION = /^{?\s*function\s+\(/i;
 
 class Parser extends EventEmitter {
     constructor(options) {
@@ -204,14 +209,6 @@ class Parser extends EventEmitter {
         }
 
         this.emit('event', item);
-    }
-
-    emitRefItem(ref) {
-        const item = Object.assign({}, ref, {
-            visibility: 'private'
-        });
-
-        this.emit('ref', item);
     }
 
     emitImportedComponentItem(component, parseContext) {
@@ -561,18 +558,11 @@ class Parser extends EventEmitter {
         this.parseBodyRecursively(ast, scriptParseContext, 0);
     }
 
-    parseMarkupExpressionBlock(expression, offset) {
+    parseTemplateJavascriptExpression(expression, offset) {
         // Add name for anonymous functions to prevent parser error
-        const regex = /^{?\s*function\s+\(/i;
-
-        expression = expression.replace(regex, function (m) {
-            // When quotes in attributes used curly braces are provided here,
-            // so we should handle it separately
-            if (m.startsWith('{')) {
-                return '{function a(';
-            }
-
-            return 'function a(';
+        expression = expression.replace(RE_ANONYMOUS_FUNCTION, function (m) {
+            // Preserve the curly brace if it appears in the quotes
+            return m.startsWith('{') ? '{function a(' : 'function a(';
         });
 
         const ast = espree.parse(
@@ -586,7 +576,7 @@ class Parser extends EventEmitter {
         });
 
         const scriptParseContext = {
-            scope: SCOPE_MARKUP,
+            // scope: SCOPE_MARKUP, // TODO: This is not used. SCOPE_TYPE instead
             offset: offset,
             sourceCode: sourceCode
         };
@@ -631,190 +621,34 @@ class Parser extends EventEmitter {
     }
 
     parseTemplate() {
-        let rootElementIndex = 0;
-        let lastComment = null;
-        let lastAttributeIndex = 0;
-        let lastAttributeLocations = {};
-        let lastTagName = null;
+        const templateParser = new TemplateParser(this);
 
-        const parser = new HtmlParser({
-            oncomment: (data) => {
-                lastComment = data.trim();
-            },
-            ontext: (text) => {
-                if (text.trim()) {
-                    lastComment = null;
-                }
-            },
-            onattribute: (name, value) => {
-                if (this.includeSourceLocations && parser.startIndex >= 0 && parser.endIndex >= parser.startIndex) {
-                    lastAttributeLocations[name] = {
-                        start: lastAttributeIndex,
-                        end: parser._tokenizer._index
-                    };
-
-                    lastAttributeIndex = parser._tokenizer._index;
-                }
-
-                if (this.features.includes('events')) {
-                    if (lastTagName !== 'slot') {
-                        // Expose events that propagated from child events
-                        // Handle event syntax like ```<button on:click>Some link</button>```
-                        if (name.length > 3 && name.indexOf('on:') === 0 && !value) {
-                            const nameWithModificators = name.substr(3).split('|');
-
-                            const baseEvent = {
-                                name: nameWithModificators[0],
-                                parent: lastTagName,
-                                modificators: nameWithModificators.slice(1),
-                                locations: this.includeSourceLocations && hasOwnProperty(lastAttributeLocations, name)
-                                    ? [lastAttributeLocations[name]]
-                                    : null
-                            };
-
-                            if (lastComment) {
-                                lastComment = `/** ${lastComment} */`;
-                            }
-
-                            const comment = utils.parseComment(lastComment || '');
-
-                            baseEvent.visibility = comment.visibility;
-                            baseEvent.description = comment.description || '';
-                            baseEvent.keywords = comment.keywords;
-
-                            if (!hasOwnProperty(this.eventsEmitted, baseEvent.name)) {
-                                this.eventsEmitted[baseEvent.name] = baseEvent;
-
-                                parseAndMergeKeywords(comment.keywords, baseEvent);
-
-                                this.emit('event', baseEvent);
-                            }
-
-                            lastComment = null;
-                        }
-
-                        // Parse event handlers
-                        if (name.length > 3 && name.indexOf('on:') === 0 && value) {
-                            this.parseMarkupExpressionBlock(value);
-                        }
-                    }
-                }
-            },
-            onopentagname: (tagName) => {
-                lastTagName = tagName;
-                lastAttributeIndex = parser._tokenizer._index;
-                lastAttributeLocations = {};
-            },
-            onopentag: (tagName, attrs) => {
-                const isNotStyleOrScript = !['style', 'script'].includes(tagName);
-                const isTopLevelElement = parser._stack.length === 1;
-
-                if (isTopLevelElement && isNotStyleOrScript) {
-                    if (lastComment && rootElementIndex === 0) {
-                        this.emitGlobalComment(utils.parseComment(lastComment));
-                    }
-
-                    rootElementIndex += 1;
-                }
-
-                if (tagName === 'slot') {
-                    if (this.features.includes('slots')) {
-                        const exposedParameters = Object.keys(attrs)
-                            .filter(name => name.length > 0 && name !== 'name')
-                            .map(name => ({
-                                name: name,
-                                visibility: 'public'
-                            }));
-
-                        const slot = {
-                            name: attrs.name || 'default',
-                            description: lastComment,
-                            visibility: 'public',
-                            parameters: exposedParameters
-                        };
-
-                        if (this.includeSourceLocations && parser.startIndex >= 0 && parser.endIndex >= parser.startIndex) {
-                            slot.loc = {
-                                start: parser.startIndex,
-                                end: parser.endIndex
-                            };
-                        }
-
-                        this.emit('slot', slot);
-                    }
-                } else {
-                    if (tagName === 'svelte:options' && attrs.tag) {
-                        if (this.features.includes('name')) {
-                            this.emit('name', attrs.tag);
-                        }
-                    }
-
-                    if (this.features.includes('data')) {
-                        const bindProperties = Object.keys(attrs)
-                            .filter(name => name.length > 5 && name.indexOf('bind:') === 0)
-                            .filter(name => name !== 'bind:this')
-                            .map(name => {
-                                const sourcePropertyName = name.substr(5);
-
-                                let targetPropertyName = sourcePropertyName;
-
-                                const attributeValue = attrs[name];
-
-                                if (attributeValue && attributeValue.length > 0) {
-                                    targetPropertyName = attributeValue;
-                                }
-
-                                return {
-                                    sourcePropertyName: sourcePropertyName,
-                                    targetPropertyName: targetPropertyName,
-                                    parent: tagName,
-                                    locations: this.includeSourceLocations && hasOwnProperty(lastAttributeLocations, name)
-                                        ? [lastAttributeLocations[name]]
-                                        : null
-                                };
-                            });
-
-                        bindProperties.forEach(bindProperty => {
-                            const dataItem = {
-                                name: bindProperty.targetPropertyName,
-                                kind: undefined,
-                                bind: [{
-                                    source: bindProperty.parent,
-                                    property: bindProperty.sourcePropertyName
-                                }],
-                                locations: bindProperty.locations,
-                                visibility: 'private',
-                                static: false,
-                                readonly: false
-                            };
-
-                            this.emit('data', dataItem);
-                        });
-                    }
-
-                    if (this.features.includes('refs')) {
-                        if (hasOwnProperty(attrs, 'bind:this') && attrs['bind:this']) {
-                            const bindedVariableName = attrs['bind:this'];
-
-                            this.emitRefItem({
-                                name: bindedVariableName,
-                                parent: tagName,
-                                locations: this.includeSourceLocations && hasOwnProperty(lastAttributeLocations, 'bind:this')
-                                    ? [lastAttributeLocations['bind:this']]
-                                    : null
-                            });
-                        }
-                    }
-                }
-            }
-        }, {
-            lowerCaseTags: false,
-            lowerCaseAttributeNames: false,
-            curlyBracesInAttributes: true
+        // Forward emit of basic template events
+        templateParser.on(TemplateEvent.DATA, (dataItem) => {
+            this.emit('data', dataItem);
+        });
+        templateParser.on(TemplateEvent.EVENT, (event) => {
+            this.emit('event', event);
+        });
+        templateParser.on(TemplateEvent.NAME, (name) => {
+            this.emit('name', name);
+        });
+        templateParser.on(TemplateEvent.REF, (refItem) => {
+            this.emit('ref', refItem);
+        });
+        templateParser.on(TemplateEvent.SLOT, (slot) => {
+            this.emit('slot', slot);
         });
 
-        parser.write(this.structure.template);
-        parser.end();
+        // Special cases where more parsing is required
+        templateParser.on(TemplateEvent.EXPRESSION, (expression) => {
+            this.parseTemplateJavascriptExpression(expression);
+        });
+        templateParser.on(TemplateEvent.GLOBAL_COMMENT, (comment) => {
+            this.emitGlobalComment(comment);
+        });
+
+        templateParser.parse();
     }
 }
